Some tidying up, added some missing cases to maths library and replaced some Allegro stuff with my own.
Also there's some logging crap in now, so that's neat.

SpriteObject no longer discards our sprite on delete, handy because it makes more sense for the actual child classes to handle this themselves.

+++ platform/platform_log.c
@@ -33,8 +33,7 @@ void plWriteLog(const char *path, const char *msg, ...)
 	vsnprintf(buffer, sizeof(buffer), msg, args);
 	va_end(args);
 
-	unsigned int size = strlen(buffer);
-
+	size_t size = strlen(buffer);
 	FILE *file = fopen(newpath, "a");
 	if (fwrite(buffer, sizeof(char), size, file) != size)
 		plSetError("Failed to write to log! (%s)", newpath);
